Move operator-only commands to the `aut protocol gov` group
